Updated to R-devel rev 72200.

+++ src/library/methods/src/methods_list_dispatch.c
@@ -783,7 +783,7 @@ static SEXP R_selectByPackage(SEXP table, SEXP classes, int nargs) {
 	thisPkg = PACKAGE_SLOT(VECTOR_ELT(classes, i));
 	if(thisPkg == R_NilValue)
 	    thisPkg = s_base;
-	lwidth += strlen(STRING_VALUE(thisPkg)) + 1;
+	lwidth += (int) strlen(STRING_VALUE(thisPkg)) + 1;
     }
     /* make the label */
     const void *vmax = vmaxget();
@@ -1024,7 +1024,7 @@ SEXP R_dispatchGeneric(SEXP fname, SEXP ev, SEXP fdef)
 		      R_curErrorBuf());
 	}
 	SET_VECTOR_ELT(classes, i, thisClass);
-	lwidth += strlen(STRING_VALUE(thisClass)) + 1;
+	lwidth += (int) strlen(STRING_VALUE(thisClass)) + 1;
     }
     /* make the label */
     const void *vmax = vmaxget();
